Create Northern Soul
# Northern Soul

I chose to make a tribute page to Northern Soul because I really like music, and Northern Soul is one of my favourite genres. It is a type of music which appeals to people of all ages, and even though most of the songs were recorded in the 1960s, the music still has wide appeal today.

##Reflections on my development:

I have **really** enjoyed learning the basics of HTML and CSS. I find it amazing that I am now able to create a basic website, and I am very excited to learn more during the remainder of the course.